Resolve SonarCloud issues with duplication
SonarCloud quite rightly highlighted the high level of duplication across the transaction file body presenter services.

Having reviewed the code we can safely move these into the base service. If we have a need to amend the functionality for a reason we can bring the impacted method out of the base then.
