+++ app/services/files/transactions/generate_presroc_transaction_file.service.js
@@ -6,27 +6,11 @@
 
 const BaseGenerateTransactionFileService = require('./base_generate_transaction_file.service')
 
-const {
-  TransactionFilePresrocBodyPresenter,
-  TransactionFileHeadPresenter,
-  TransactionFileTailPresenter
-} = require('../../../presenters')
+const TransactionFilePresrocBodyPresenter = require('../../../presenters/transaction_file_presroc_body.presenter')
 
 class GeneratePresrocTransactionFileService extends BaseGenerateTransactionFileService {
-  static _headPresenter () {
-    return TransactionFileHeadPresenter
-  }
-
   static _bodyPresenter () {
     return TransactionFilePresrocBodyPresenter
-  }
-
-  static _tailPresenter () {
-    return TransactionFileTailPresenter
-  }
-
-  static _join () {
-    return ['invoices', 'transactions.invoiceId', 'invoices.id']
   }
 
   static _select () {
@@ -56,32 +40,6 @@ class GeneratePresrocTransactionFileService extends BaseGenerateTransactionFileS
       'invoices.creditLineValue',
       'invoices.debitLineValue'
     ]
-  }
-
-  static _sort () {
-    return [
-      'invoices.transactionReference', // sort by transaction reference
-      'lineAttr1', // then sort by licence number
-      'regimeValue17' // then sort by compensation charge, where non-compensation charge (ie. false) is first
-    ]
-  }
-
-  static _where (builder, billRun) {
-    return builder
-      .where('transactions.billRunId', billRun.id)
-      .whereNotNull('invoices.transactionReference')
-      .whereNot('chargeValue', 0)
-  }
-
-  static _additionalData (billRun) {
-    return {
-      region: billRun.region,
-      billRunNumber: billRun.billRunNumber,
-      billRunUpdatedAt: billRun.updatedAt,
-      invoiceValue: billRun.invoiceValue,
-      creditNoteValue: billRun.creditNoteValue,
-      fileReference: billRun.fileReference
-    }
   }
 }
